requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
pandas
numpy
scikit-learn
xgboost
